refactoring event in progress (/events)

+++ main-service/src/main/java/ru/practicum/ewm/user/service/UserServiceImpl.java
@@ -3,6 +3,7 @@ package ru.practicum.ewm.user.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ru.practicum.ewm.exception.NotFoundException;
 import ru.practicum.ewm.user.dto.UserCreateDto;
 import ru.practicum.ewm.user.dto.UserDto;
 import ru.practicum.ewm.user.mapper.UserMapper;
@@ -35,6 +36,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new NotFoundException(String.format("User with id=%d was not found", id));
+        }
         repository.deleteById(id);
     }
 }
